ensured that data.hdr.FirstTimeStamp is converted to double in ft_spike_maketrials.m
git-svn-id: svn+ssh://svn.fcdonders.nl/home/svnroot/fieldtrip/trunk@6058 0cf7c7f0-3615-4144-b4e6-68da3bce3cd0

+++ contrib/spike/ft_spike_maketrials.m
@@ -141,8 +141,8 @@ else
   nTrials = size(trl,1);
   
   try
-    FirstTimeStamp     = data.hdr.FirstTimeStamp;
-    TimeStampPerSample = data.hdr.TimeStampPerSample;
+    FirstTimeStamp     = double(data.hdr.FirstTimeStamp);
+    TimeStampPerSample = double(data.hdr.TimeStampPerSample);
   catch
     error('could not find the timestamp information in the continuous data');
   end
